+++ examples/forest_fire.jl
@@ -23,7 +23,7 @@ gr() # hide
 
 mutable struct Tree <: AbstractAgent
     id::Int
-    pos::Tuple{Int,Int}
+    pos::Dims{2}
     status::Symbol  #:green, :burning, :burnt
 end
 nothing # hide
@@ -86,7 +86,7 @@ forest = forest_fire(griddims = (20, 20))
 burnt_percentage(m) = count(t->t.status == :burnt, allagents(m)) / length(positions(m))
 mdata = [burnt_percentage]
 
-_, data = run!(forest, tree_step!, 10; mdata = mdata)
+_, data = run!(forest, tree_step!, 10; mdata)
 data
 
 # Now let's plot the model. We use green for unburnt trees, red for burning and a