+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Performance increase of finding neighbors in GraphSpace with r > 1.
 - New wrapping function `nearby_agents` that returns an iterable of neighboring agents.
 - Positions and neighbors on `GridSpace` can now be searched in each direction separately by accepting `r` as a tuple.
+- Neighbors on non-periodic chebyshev spaces can also be searched per dimension over a specific range.
 - New public `schedule` function for writing custom loops.
 - Mixed models are supported in data collection methods.
 - `random_agent(model, condition)` allows obtaining random agents that satisfy given condition.
@@ -35,6 +36,9 @@ therefore are not "truly breaking".
   - `find_empty_nodes` -> `empty_positions`
   - `has_empty_nodes` -> `has_empty_positions`
   - `nodes` -> `positions`
+
+## Non-breaking changes
+- `GridSpace` agents now use `Dims` rather than `Tuple{N,Int}` for their `pos`ition in all examples and pre-defined models.
 
 # v3.7
 - Add the ability to decide whether the agent step or the model step should be performed first using the `agents_first` argument.